timer is component now
